Extracts scrambleHex to composable
