+++ frontend/app/.eslintrc-auto-import.json
@@ -282,6 +282,7 @@
     "useRound": true,
     "useRoute": true,
     "useRouter": true,
+    "useScramble": true,
     "useScreenOrientation": true,
     "useScreenSafeArea": true,
     "useScriptTag": true,
